feat: 공통 component common 폴더로 빼기
