+++ src/AnimatedBunney/index.ts
@@ -1,0 +1,2 @@
+export { AnimatedBunney } from './AnimatedBunney'
+export type { BunneyProps as AnimatedBunneyProps } from './AnimatedBunney'
