+++ src/Bunney/Bunney.tsx
@@ -1,29 +1,8 @@
-import * as S from './styled'
-import StartIcon from '../../public/Start.svg?react'
-import { WholeBunney } from '@/common/styled'
+import { BunneyProps as CommmonBunneyProps, Bunney as CommonBunney } from '@/common/Bunney'
 
-export interface BunneyProps {
-    color?: string
-    earColor?: string
-    eyeColor?: string
-}
+export interface BunneyProps extends CommmonBunneyProps{}
 export const Bunney = ({ color, earColor, eyeColor }: BunneyProps) => {
     return (
-        <WholeBunney>
-            <S.EarWrapper>
-                <S.Ear dir="left" earColor={earColor} />
-                <S.Ear dir="right" earColor={earColor} />
-            </S.EarWrapper>
-            <S.FaceWrapper color={color}>
-                <S.EyeWrapper>
-                    <S.Eye eyeColor={eyeColor}>
-                        <StartIcon />
-                    </S.Eye>
-                    <S.Eye eyeColor={eyeColor}>
-                        <StartIcon />
-                    </S.Eye>
-                </S.EyeWrapper>
-            </S.FaceWrapper>
-        </WholeBunney>
+        <CommonBunney color={color} earColor={earColor} eyeColor={eyeColor} />
     )
 }
