+++ src/Mackey/Mackey.tsx
@@ -1,13 +1,8 @@
-import { Logo } from '../Logo';
-import * as S from './styled'
-export interface MackeyProps{
-    size?: number;
-    color?: string;
+import {Mackey as CommonMackey,MackeyProps as MackeyCommonProps} from '@/common/Mackey'
+export interface MackeyProps extends MackeyCommonProps{
 }
 export const Mackey = ({size,color}:MackeyProps) => {
     return (
-        <S.Mac size={size}>
-            <Logo size={size ? size / 7 : 1.35} color={color || 'linear-gradient(#ff001f 19%,#e511aa 52%,#c03eed 98%)'} />
-        </S.Mac>
+        <CommonMackey size={size} color={color} />
     )
 }
